basic skeleton of Detail Componentfor storybook manager
