+++ src/components/pages/StorybookManager/AudiobookList.js
@@ -2,36 +2,7 @@ import React, { useState, useEffect } from 'react';
 import { Skeleton, Table, Space, Button } from 'antd';
 import { useStore } from './Context/StorybookManager.Context';
 import { showModalAddAudiobook } from './Context/StorybookManager.Actions';
-
-const sampleData = [
-  {
-    id: 1,
-    trackName: "Harry Potter and the Philosopher's Stone - Chapter 1",
-    length: '10:29',
-    chapter: 1,
-    episode: 1,
-    author: 'J. K. Rowling',
-    story: "Harry Potter and the Philosopher's Stone",
-  },
-  {
-    id: 2,
-    trackName: "Harry Potter and the Philosopher's Stone - Chapter 2",
-    length: '15:29',
-    chapter: 2,
-    episode: 1,
-    author: 'J. K. Rowling',
-    story: "Harry Potter and the Philosopher's Stone",
-  },
-  {
-    id: 3,
-    trackName: "Harry Potter and the Philosopher's Stone - Chapter 3",
-    length: '30:29',
-    chapter: 3,
-    episode: 2,
-    author: 'J. K. Rowling',
-    story: "Harry Potter and the Philosopher's Stone",
-  },
-];
+import sampleData from './SampleData';
 
 const createTableColumns = render => {
   return [
